Address review feedback: split post preview cards per tool, drop redundant ?? undefined
