+++ webapp/src/components/llmbot_post/turn_content_utils.ts
@@ -102,9 +102,9 @@ function toolUseBlockToToolCall(block: ContentBlock, resultMap: Map<string, Cont
         name: block.name ?? '',
 
         description: block.description ?? '',
-        title: block.title ?? undefined, // eslint-disable-line no-undefined
-        server_origin: block.server_origin ?? undefined, // eslint-disable-line no-undefined
-        mcp_bare_name: block.mcp_bare_name ?? undefined, // eslint-disable-line no-undefined
+        title: block.title,
+        server_origin: block.server_origin,
+        mcp_bare_name: block.mcp_bare_name,
         arguments: (block.input as ToolCall['arguments']) ?? undefined, // eslint-disable-line no-undefined
         result: resultBlock?.content ?? undefined, // eslint-disable-line no-undefined
         status: statusStringToEnum(block.status),
